[fp522] feat(test): pipe input, correct input and output strings, do not run 'advanced' tests

+++ src/test/Generator.scala
@@ -4,9 +4,10 @@ import src.main.wacc._
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers._
 import org.scalatest.prop.TableDrivenPropertyChecks
+
 import scala.sys.process._
 import scala.io.Source
-import java.io.{File, FileOutputStream, PrintStream}
+import java.io.{BufferedWriter, File, FileOutputStream, FileWriter, PrintStream}
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.{Await, Future}
 import scala.concurrent.duration.DurationInt
@@ -18,23 +19,26 @@ class Generator extends AnyFlatSpec with TableDrivenPropertyChecks {
 
   behavior of "generator"
   forAll(Table("cases", TestFiles("valid/"): _*)) { file =>
-    it should s"produce the correct output and exit code for ${file.getParentFile}/${file.getName}" in {
-      val filePath = file.getPath
-      val filename = file.getName.replaceFirst("\\.wacc$", "")
-      val nullPrintStream = new PrintStream(new FileOutputStream("/dev/null"))
-      Console.withOut(nullPrintStream) {
-        Main.main(Array(filePath))
-      }
-      compileAssembly(s"$filename.s") shouldEqual None
-      val (input, expOut, expExit) = parseWaccFile(filePath)
-      try {
-        val (out, exit) = Await.result(Future(runBinary(filename, input)), 60.seconds)
+    if (file.getParentFile.getName != "advanced") {
+      it should s"produce the correct output and exit code for ${file.getParentFile}/${file.getName}" in {
+        val filePath = file.getPath
+        val filename = file.getName.replaceFirst("\\.wacc$", "")
+        Console.withOut(new PrintStream(new FileOutputStream("/dev/null"))) {
+          Main.main(Array(filePath))
+        }
+        compileAssembly(s"$filename.s") shouldEqual None
+        val (input, expOut, expExit) = parseWaccFile(filePath)
+        var ret = ("", 1)
+        try {
+          ret = Await.result(Future(runBinary(filename, input)), 60.seconds)
+        } catch {
+          case e: Exception =>
+            deleteFile(filename)
+            fail(s"Error running binary: ${e.getMessage}")
+        }
+        val (out, exit) = ret
         out shouldEqual expOut
         exit shouldEqual expExit
-      } catch {
-        case e: Exception =>
-          deleteFile(filename)
-          fail(s"Error running binary: ${e.getMessage}")
       }
     }
   }
@@ -54,7 +58,7 @@ class Generator extends AnyFlatSpec with TableDrivenPropertyChecks {
   }
 
   def runBinary(binaryFile: String, input: String): (String, Int) = {
-    val process = s"./$binaryFile"
+    val process = s"echo $input" #> s"./$binaryFile"
     new File(binaryFile).setExecutable(true)
     val output = new StringBuilder
     val exitCode = process ! ProcessLogger(output.append(_))
@@ -71,11 +75,11 @@ class Generator extends AnyFlatSpec with TableDrivenPropertyChecks {
 
     for (line <- lines) {
       if (line.startsWith("# Output:")) {
-        output = lines.takeWhile(!_.isBlank).map(_.drop(2)).mkString("\n")
+        output = lines.takeWhile(!_.isBlank).map(_.drop(2)).mkString("")
       } else if (line.startsWith("# Exit:")) {
         exitCode = lines.next().drop(2).toInt
       } else if (line.startsWith("# Input:")) {
-        input = lines.takeWhile(!_.isBlank).map(_.drop(2)).mkString("\n")
+        input = lines.takeWhile(!_.isBlank).map(_.drop(2)).mkString("")
       }
     }
